Analisis 3 y base de analisis y reportes

+++ IReporte.cpp
@@ -7,7 +7,7 @@
 // Código de apoyo....
 int IReporte:: AplicacionADNcalcularConcentracion(const string& textoBuscar, const string& textoBase) {
     int repeticiones = 0;
-    unsigned long index = 0; // la posición donde se encuentra la letra
+    size_t  index = 0; // la posición donde se encuentra la letra
     // El siguiente ciclo permite recorrer el string para identificar las veces que se
     // repite un carácter
     while ((index = textoBase.find(textoBuscar, index)) != std::string::npos) {
